move fixtures to closest conftest
